+++ src/pages/home/homePage.tsx
@@ -1,7 +1,7 @@
 import './homePage.css';
 import { Col, Container, Row } from 'react-bootstrap';
-import { SwitchPages5 } from './interfaces/SwitchPages';
-import { SwitchPage } from './components/SwitchPage';
+import { SwitchPages5 } from '../../interfaces/SwitchPages';
+import { SwitchPage } from '../../components/SwitchPage';
 //import { relative } from 'path'; 
 
 export function HomePage ({setCurrentPage, blurPage, setBlurPage}: SwitchPages5) {
